+++ include/json.hxx
@@ -86,13 +86,20 @@ namespace json {
 		data(double number) { operator=(number); }
 		data& operator=(double number) { set(number); return *this; }
 
-		operator bool() const { validate(_Type::Boolean); return _value == grammar::_valueTrue; }
-		operator int() const { validate(_Type::Number); return std::atoi(_value.c_str()); }
-		operator double() const { validate(_Type::Number); return std::atof(_value.c_str()); }
+		bool toBool() const { validate(_Type::Boolean); return _value == grammar::_valueTrue; }
+		int toInt() const { validate(_Type::Number); return std::atoi(_value.c_str()); }
+		double toDouble() const { validate(_Type::Number); return std::atof(_value.c_str()); }
+		std::string toString() const {
+			validate(_Type::String);
+			return _value;
+		}
+
+		operator bool() const { return toBool(); }
+		operator int() const { return toInt(); }
+		operator double() const { return toDouble(); }
 		operator const char* () const {
 			if (_type == _Type::Null) return nullptr;
-			validate(_Type::String);
-			return _value.c_str();
+			return toString().c_str(); 
 		}
 		operator std::vector<data>() const { validate(_Type::Array); return _elements; }